spark/requirements.txt
delta-spark
pymongo